+++ src/modules/claim-link/types/initialize/index.ts
@@ -1,0 +1,3 @@
+type TInitialize= () => Promise<void> // txhash
+
+export default TInitialize